skills added

+++ src/functions/participant/DN_team/addSkills.js
@@ -1,10 +1,22 @@
+const errorHandler = require('../../../middleware/errorHandler')
 const SkillVacancy = require('../../../models/SkillVacancy')
+const { BadRequestError } = require('../../../utils/error')
 
 //trycatch
 
 const addSkills = async(req,res) =>{
+    try {
+        const skills = req.body.skills
+    if(skills.length==0){
+        return res.status(404).send('Please enter some skills')
+    }
+    const allowed_skills = ['frontend','backend','ml','ui/ux','appdev','management','blockchain','cybersecurity']
+    const isAllowed = skills.every((skill)=>allowed_skills.includes(skill))
+    console.log(isAllowed)
+    if(!isAllowed){
+        return res.status(403).send('Invalid skills')
+    }
     await SkillVacancy.deleteMany({team_id:req.params.team_id})
-    const skills = req.body.skills
         let skillRecords = []
         let i = 0
         skills.forEach(async(skill) => {
@@ -19,6 +31,10 @@ const addSkills = async(req,res) =>{
                 res.status(201).send(skillRecords)
             }
         })
+    } catch (e) {
+        errorHandler(new BadRequestError,req,res)
+    }
+    
 }
 
 module.exports = addSkills
